fix: rename lambda function
- rename lambda function
- fix jsdoc
- add loggin
